exercises 4.15-4.21 complete, app backend working

+++ partFour/blogList/controllers/blogs.js
@@ -15,6 +15,18 @@ blogRouter.get("/", async (request, response, next) => {
   }
 });
 
+blogRouter.get("/:id", async (req, res, next) => {
+  try {
+    const blog = await Blog.findById(req.params.id).populate("user", {
+      username: 1,
+      name: 1,
+    });
+    res.json(blog);
+  } catch (error) {
+    next(error);
+  }
+});
+
 blogRouter.post("/", async (request, response, next) => {
   try {
     const body = request.body;
@@ -46,8 +58,25 @@ blogRouter.post("/", async (request, response, next) => {
 });
 
 blogRouter.delete("/:id", async (request, response, next) => {
+  // we need to decode the current token and get the current user id
+  // and we need to get the user id of the blog to be deleted
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    const blogToDelete = await Blog.findById(request.params.id);
+
+    if (!decodedToken.id)
+      return response.status(401).json({ error: "no token found" });
+    if (decodedToken.id !== blogToDelete.user.toString()) {
+      return response
+        .status(401)
+        .json({ error: "only blog creator can delete blog" });
+    }
+
+    const authorOfBlog = await User.findById(blogToDelete.user.toString());
+    authorOfBlog.blogs = authorOfBlog.blogs.pop(blogToDelete.id);
+    await authorOfBlog.save();
     await Blog.findByIdAndDelete(request.params.id);
+
     response.status(204).end();
   } catch (error) {
     next(error);
